I hate opencv
