+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/Field.java
@@ -6,21 +6,7 @@ import org.opencv.core.Point3;
 public class Field {
     //in
     //left to right
-    final static Point3 APIRL_TAG_BOARD_POSTIONS[] = {
-            new Point3(  60.25f, 41.41f,  4f),
-            new Point3( 60.25f,  35.41f,  4f),
-            new Point3(  60.25f,  29.41f,  4f),
-            new Point3(  60.25f,  -29.41f,  4f),
-            new Point3(  60.25f,  -35.41f,  4f),
-            new Point3(  60.25f,  -41.41f,  4f),
-            };
-    //Bigs are the 5.5s and 4s are the small
-    final static Point3 APIRL_TAG_WALL_POSTIONS[] = {
-            new Point3(  -70.25f,  -40.625f,  5.5f),
-            new Point3(  -70.25f,  -35.125f,  4f),
-            new Point3(  -70.25f,  35.125f,  4f),
-            new Point3(  -70.25f,  40.625f,  5.5f),
-    };
+
 
     final static double TILE_SIZE = 24;
     final static double TILE_CONNECTOR_SIZE = 1;
@@ -48,13 +34,7 @@ public class Field {
     }
     //not offset
     static Pixel[][] pixels = new Pixel[PIXEL_COLUMN][PIXEL_ROW];
-    public static Point3 getTagPosition(int i) {
-        if(i>APIRL_TAG_BOARD_POSTIONS.length-1) {
-            return APIRL_TAG_WALL_POSTIONS[i];
-        } else {
-            return APIRL_TAG_BOARD_POSTIONS[i];
-        }
-    }
+
 
 
 }
